+++ backend/src/main/java/nl/uva/alexandria/model/sattose/ExternalCall.java
@@ -3,17 +3,17 @@ package nl.uva.alexandria.model.sattose;
 import java.util.Map;
 
 public class ExternalCall {
-    private Integer sourceMethod;
+    private Long sourceMethod;
     private String targetMethod;
     private Map<String, String> metadata;
 
-    public ExternalCall(Integer sourceMethod, String targetMethod, Map<String, String> metadata) {
+    public ExternalCall(Long sourceMethod, String targetMethod, Map<String, String> metadata) {
         this.sourceMethod = sourceMethod;
         this.targetMethod = targetMethod;
         this.metadata = metadata;
     }
 
-    public Integer getSourceMethod() {
+    public Long getSourceMethod() {
         return sourceMethod;
     }
